// RMG/ClientApp/src/app/models/Assesment.ts
export class Assesment {
  assessment_type: string;
  mode_of_interview: string;
  rating: string;
  interviwed_by: string;
  remarks: string;
  assessment_date: string;
  promoted: string;

  applicant_code: string;
  applicant_name: string;
  educational_details: string;
  experience: string;
  previous_company: string;
  employment_type: string;
  skills: string;
  phone: string;
  email_id: string;
}

//applicant_id,
//  asmt_round,
//  mode_of_interview,
//  assessment_type,
//  skillset,
//  rating,
//  interviwed_by,
//  remarks,
//  assessment_date,
//  promoted